requirements.txt
ruamel.yaml >= 0.16
sevenbridges-python >= 2.0
nf-core==2.1
cwlformat
packaging
